feat(demos): per-demo model switcher via shared switch_api router
Move the model-switch API (/api/status, /api/profiles, /api/switch) into a
shared common/backend/switch_api.py and mount it on every demo backend, so each
demo page carries its own switcher (kind-limited: asr-caption/diarization → ASR,
tts-playground/voice-clone → TTS, v2v-chat → both) alongside the gallery's.
createModelSwitchPanel gains opts.kinds — a single kind hides the ASR/TTS toggle.
Gallery refactored to import the shared router (behavior unchanged).

Also fixes a latent crash the extraction surfaced: /api/profiles 500'd when the
SLV was unreachable (backend_status is None); guard it so a demo with the voice
server down still returns the platform listing. Test updated: a TTS-only profile
(kokoro-trt) no longer appears in the ASR list — the model-dedup layer correctly
drops profiles that declare no backend for the requested kind.
