+++ demos/tts-playground/backend/main.py
@@ -49,8 +49,10 @@ if str(_DEMOS_DIR) not in sys.path:
     sys.path.insert(0, str(_DEMOS_DIR))
 
 from common.backend.slv_proxy import SLVProxy  # noqa: E402
+from common.backend.switch_api import register_switch_routes  # noqa: E402
 
 _TRUTHY = {"1", "true", "yes", "on"}
+_DEFAULT_PROFILES_DIR = _DEMOS_DIR.parent / "configs" / "profiles"
 
 
 class TTSProxyRequest(BaseModel):
@@ -145,6 +147,13 @@ def create_app(proxy: SLVProxy | None = None, kiosk: bool | None = None) -> Fast
     app.post("/api/tts/stream")(_tts_stream)
     # Alias for the shared TTSStreamPlayer, which POSTs {origin}/tts/stream.
     app.post("/tts/stream")(_tts_stream)
+
+    # Shared model-switch API (/api/status, /api/profiles, /api/switch). This
+    # demo's frontend panel is pinned to ["tts"]. Registered BEFORE the mount.
+    profiles_dir = Path(os.environ.get("DEMO_PROFILES_DIR") or _DEFAULT_PROFILES_DIR)
+    register_switch_routes(
+        app, slv, profiles_dir, asr_label=os.environ.get("DEMO_ASR_MODEL_ID")
+    )
 
     # Static frontend (mounted last so /api/* wins).
     common_frontend = _DEMOS_DIR / "common" / "frontend"
