+++ demos/v2v-chat/backend/main.py
@@ -24,6 +24,7 @@ from __future__ import annotations
 
 import contextlib
 import os
+import sys
 from pathlib import Path
 from urllib.parse import urlsplit
 
@@ -34,7 +35,16 @@ _HERE = Path(__file__).resolve().parent          # demos/v2v-chat/backend
 _APP_DIR = _HERE.parent                           # demos/v2v-chat
 _DEMOS_DIR = _APP_DIR.parent                      # demos/
 
+# Allow running as a plain script (`uv run python v2v-chat/backend/main.py`):
+# the demos/ dir must be importable for common.backend.* .
+if str(_DEMOS_DIR) not in sys.path:
+    sys.path.insert(0, str(_DEMOS_DIR))
+
+from common.backend.slv_proxy import SLVProxy  # noqa: E402
+from common.backend.switch_api import register_switch_routes  # noqa: E402
+
 DEFAULT_SLV_URL = "http://127.0.0.1:8621"
+_DEFAULT_PROFILES_DIR = _DEMOS_DIR.parent / "configs" / "profiles"
 _LOOPBACK_HOSTS = {"127.0.0.1", "localhost", "::1", "0.0.0.0"}
 
 
@@ -56,18 +66,20 @@ def _ws_endpoint(slv_url: str) -> dict:
     }
 
 
-def create_app(slv_url: str | None = None) -> FastAPI:
+def create_app(slv_url: str | None = None, proxy: SLVProxy | None = None) -> FastAPI:
     slv_url = (slv_url or os.environ.get("SLV_URL") or DEFAULT_SLV_URL).rstrip("/")
+    # The model-switch panel proxies SLV admin routes through this backend; the
+    # WS voice-chat path still hits SLV directly from the browser.
+    slv = proxy or SLVProxy(base_url=slv_url)
 
     @contextlib.asynccontextmanager
     async def _lifespan(_app: FastAPI):
-        # No pooled resources yet (the browser connects to SLV directly);
-        # kept for parity with the gallery app factory so future proxy
-        # clients get a teardown home.
         yield
+        await slv.aclose()
 
     app = FastAPI(title="slv-demo-v2v-chat", docs_url=None, redoc_url=None,
                   lifespan=_lifespan)
+    profiles_dir = Path(os.environ.get("DEMO_PROFILES_DIR") or _DEFAULT_PROFILES_DIR)
 
     @app.get("/healthz")
     async def healthz() -> dict:
@@ -80,6 +92,13 @@ def create_app(slv_url: str | None = None) -> FastAPI:
             "ws": _ws_endpoint(slv_url),
             "v2v_path": "/v2v/stream",
         }
+
+    # Shared model-switch API (/api/status, /api/profiles, /api/switch). This
+    # demo runs the full ASR→LLM→TTS loop, so its frontend panel exposes both
+    # ["asr","tts"] tabs. Registered BEFORE the static mount.
+    register_switch_routes(
+        app, slv, profiles_dir, asr_label=os.environ.get("DEMO_ASR_MODEL_ID")
+    )
 
     # Static frontend (mounted last so /api/* and /healthz win).
     common_frontend = _DEMOS_DIR / "common" / "frontend"
