+++ demos/v2v-chat/frontend/index.html
@@ -279,6 +279,11 @@
     </section>
 
     <section class="panel">
+      <h2 data-i18n="switchTitle"></h2>
+      <div id="switch-panel"></div>
+    </section>
+
+    <section class="panel">
       <h2 data-i18n="settingsTitle"></h2>
       <label class="muted" style="font-size: var(--fs-small);" for="asr-lang" data-i18n="asrLangLabel"></label>
       <select id="asr-lang" style="width: 100%; margin-top: 6px; margin-bottom: 12px;">
