+++ demos/voice-clone/frontend/index.html
@@ -340,6 +340,10 @@
     <section class="panel">
       <p class="muted" style="font-size:var(--fs-small); margin:0" data-i18n="aboutClone"></p>
     </section>
+    <section class="panel">
+      <h2 data-i18n="switchTitle"></h2>
+      <div id="switch-panel"></div>
+    </section>
   </aside>
 </div>
